cleaning out r-func.ts
